# utils/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/users/check": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users/sign-up": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users/login": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users/password": {
      "put": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/users/profile": {
      "patch": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/upload/image": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/portfolio/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/portfolio/{cardId}/canvas": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/portfolio/{cardId}/publish": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/bookmark-list/cards/{cardId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/bookmark-list/cards/{cardId}/pin": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/bookmark-list/cards/{cardId}/notes": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/bookmark-list/groups": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/bookmark-list/groups/{followerGroupId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "followerGroupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "followerGroupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/bookmark-list/groups/order": {
      "patch": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/api/bookmark-list/groups/{followerGroupId}/cards": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "followerGroupId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/homepage/{cardId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/homepage/{cardId}/page-title": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/homepage/{cardId}/link": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/homepage/{cardId}/link/{linkId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "failOnErrors": true,
  "definition": {
    "openapi": "3.0.0",
    "info": {
      "title": "MEISHIcian API"
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "value": "Bearer <JWT token here>"
        }
      }
    },
    "tags": [
      {
        "name": "1. 會員驗證",
        "description": "會員驗證"
      },
      {
        "name": "2. 名片管理",
        "description": "名片管理相關功能"
      },
      {
        "name": "3. 個人名片",
        "description": "名片管理相關功能"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ]
  },
  "apis": [
    "./routes/*.js",
    "./validations/*.js"
  ]
}